+++ libs/adls/tests/test_adls/test_resource.py
@@ -1,10 +1,18 @@
+import os
+import random
 from pathlib import Path
+from unittest import mock
 
+import pytest
+
+from thds.adls import resource
 from thds.adls.fqn import AdlsFqn
-from thds.adls.resource import AdlsHashedResource, upload
+from thds.adls.resource import AdlsHashedResource, get_read_only, upload, verify_or_create
+from thds.adls.resource.file_pointers import MustCommitResourceLocally
 from thds.adls.ro_cache import global_cache
 
-HW = Path(__file__).parent.parent / "data/hello_world.txt"
+DATA_DIR = Path(__file__).parent.parent / "data"
+HW = DATA_DIR / "hello_world.txt"
 
 
 def test_serde():
@@ -58,3 +66,46 @@ def test_upload_iterable_bytes():
         (b"87987987987987897987987987987987987", b"45345435453"),
         write_through_cache=global_cache(),
     )
+
+
+@mock.patch.dict(os.environ, {"CI": ""})
+def test_get_or_create_resource():
+    the_path = DATA_DIR / "delete_me.db"
+    the_path.unlink(missing_ok=True)
+    adls_pointer = DATA_DIR / "does-not-exist.adls"
+    adls_pointer.unlink(missing_ok=True)
+
+    def describe():
+        return AdlsFqn.parse("adls://thdsscratch/tmp/test/delete_me.db")
+
+    creations = 0
+
+    def create():
+        nonlocal creations
+        creations += 1
+        with open(the_path, "w") as f:
+            f.write(str(random.random()) * 10_000)
+        return the_path
+
+    verify_or_create(adls_pointer, describe, create)
+    res = verify_or_create(adls_pointer, describe, create)
+    # second time does no creation.
+    assert creations == 1
+
+    assert the_path.exists()
+    get_read_only(res, the_path)
+    assert the_path.exists()
+    the_path.unlink()
+    adls_pointer.unlink()
+
+
+@mock.patch.dict(os.environ, {"CI": "1"})
+def test_in_ci_resource_to_path_fails():
+    with pytest.raises(MustCommitResourceLocally):
+        resource.to_path(
+            "whatever.txt",
+            AdlsHashedResource.parse(
+                '{"uri": "adls://foo/bar/baz", "md5b64": "WPMVPiXYwhMrMjF87w3GvA=="}'
+            ),
+            check_ci=True,
+        )